Eliminate Columns class
I only introduced this class recently, breaking some code out of UIStarter.  Then UIStarter (mostly) became ResultsTable, and now I think actually the code works just fine inside ResultsTable.
